Increase size of drawer and theme toggle icons
- Refactor drawer and code icons to seperate components
- Refactor footer to seperate component
- Refactor desktop and mobile menu to seperate components.
