+++ src/App.tsx
@@ -1,3 +1,4 @@
+import Footer from "./components/footer/Footer";
 import Intro from "./components/intro/intro";
 import Layout from "./components/layout/layout";
 import Navbar from "./components/navbar/navbar";
@@ -17,7 +18,7 @@ const App = () => {
           <div className="flex flex-col">
             <Intro />
           </div>
-          <div>
+          <div className="pb-10">
             {siteSections.map((section) => {
               const sectionData = siteData[section];
               return (
@@ -32,29 +33,7 @@ const App = () => {
               );
             })}
           </div>
-          <div className="flex flex-col justify-center mb-4 mt-4">
-            <div className="text-secondary text-center opacity-70 text-sm mb-2">
-              Designed and developed by Thomas Rainford
-            </div>
-            <div className="flex flex-col justify-center items-center">
-              <a href={"https://github.com/ThomasRainford"} target="_blank">
-                <i className="devicon-github-original text-lg text-secondary opacity-70"></i>
-              </a>
-            </div>
-          </div>
-          <div className="flex justify-center mb-4">
-            <button
-              className="btn btn-soft btn-secondary px-2 py-1"
-              onClick={() => {
-                window.scrollTo({
-                  top: 0,
-                  behavior: "smooth",
-                });
-              }}
-            >
-              Scroll To Top
-            </button>
-          </div>
+          <Footer />
         </div>
       </Layout>
     </main>
