+++ src/components/icons/SunIcon.tsx
@@ -1,7 +1,7 @@
 const SunIcon = () => {
   return (
     <svg
-      className="swap-off h-7 w-7 fill-current"
+      className="swap-off h-9 w-9 fill-current"
       xmlns="http://www.w3.org/2000/svg"
       viewBox="0 0 24 24"
     >
