+++ src/components/intro/intro.tsx
@@ -1,3 +1,5 @@
+import CodeIcon from "../icons/CodeIcon";
+
 const Intro = () => {
   return (
     <div className="flex flex-col mt-15 h-screen">
@@ -21,7 +23,7 @@ const Intro = () => {
           I build scalable and maintainable soultions with great user
           experiences. Currently working at{" "}
           <a
-            className="text-base-content hover:underline"
+            className="text-base-content underline hover:text-accent"
             target="_blank"
             href="https://hortplus.com"
           >
@@ -31,25 +33,7 @@ const Intro = () => {
         </p>
       </div>
       <div className="mt-5 pl-4">
-        <svg
-          className="w-[200px] h-[200px] sm:w-[250px] sm:h-[250px]"
-          xmlns="http://www.w3.org/2000/svg"
-          viewBox="0 0 24 24"
-          fill="none"
-        >
-          <path
-            d="M14.9615 5.27473C15.1132 4.7437 14.8058 4.19021 14.2747 4.03849C13.7437 3.88677 13.1902 4.19426 13.0385 4.72529L9.03847 18.7253C8.88675 19.2563 9.19424 19.8098 9.72528 19.9615C10.2563 20.1133 10.8098 19.8058 10.9615 19.2747L14.9615 5.27473Z"
-            className="fill-neutral"
-          />
-          <path
-            d="M5.7991 7.39879C6.13114 7.84012 6.04255 8.46705 5.60123 8.7991L2.40894 11.2009C1.87724 11.601 1.87723 12.399 2.40894 12.7991L5.60123 15.2009C6.04255 15.533 6.13114 16.1599 5.7991 16.6012C5.46705 17.0426 4.84012 17.1311 4.39879 16.7991L1.20651 14.3973C-0.388615 13.1971 -0.388621 10.8029 1.2065 9.60276L4.39879 7.20093C4.84012 6.86889 5.46705 6.95747 5.7991 7.39879Z"
-            className="fill-neutral"
-          />
-          <path
-            d="M18.2009 16.6012C17.8689 16.1599 17.9575 15.533 18.3988 15.2009L21.5911 12.7991C22.1228 12.399 22.1228 11.601 21.5911 11.2009L18.3988 8.7991C17.9575 8.46705 17.8689 7.84012 18.2009 7.39879C18.533 6.95747 19.1599 6.86889 19.6012 7.20093L22.7935 9.60276C24.3886 10.8029 24.3886 13.1971 22.7935 14.3973L19.6012 16.7991C19.1599 17.1311 18.533 17.0426 18.2009 16.6012Z"
-            className="fill-neutral"
-          />
-        </svg>
+        <CodeIcon />
       </div>
     </div>
   );
